@@ -1,7 +1,1 @@
-'use strict';
-
-module.exports = module.exports.Transformer = require('./lib/transformer.js');
-module.exports.Reader = require('./lib/reader.js');
-module.exports.Writer = require('./lib/writer.js');
-module.exports.middleware = require('./lib/middleware.js');
-module.exports.constants = require('./lib/constants.js');
+module.exports = require('./lib/jy-transform');
